move 2 methods back to cli_conf
